+++ .gitignore
@@ -1,2 +1,6 @@
 __pycache__/
 .pytest_cache/
+venv*/
+.coverage
+coverage.xml
+benchmark*
